Move category components to (main)/category/components
