+++ client/app/(main)/category/components/Categories.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import React from 'react'
 import { CategoryItem } from './CategoryItem'
 import { Category } from '@/app/types'
