+++ client/app/(main)/category/components/CategorySection.tsx
