+++ client/app/(main)/category/components/index.ts
@@ -1,3 +1,5 @@
+'use client'
+
 export * from './CategorySection'
 export * from './CreateCategory'
 export * from './CategoryDisplay'
