+++ client/app/(main)/category/layout.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { Container } from '@/app/ui'
+
+export default function layout({
+  children
+}: Readonly<{
+  children: React.ReactNode
+}>) {
+  return <Container>{children}</Container>
+}
