+++ client/app/(main)/category/page.tsx
@@ -1,16 +1,15 @@
-import { Container, CategorySection, Title, CreateCategory, CategoryDisplay } from '@/app/ui'
+import { Title } from '@/app/ui'
 import { getCategoryList } from '@/app/utils'
+import { CategoryDisplay, CategorySection, CreateCategory } from '@/app/(main)/category/components'
 
 export default async function page() {
   const { data } = await getCategoryList()
 
   return (
-    <Container>
-      <CategorySection>
-        <Title>{String('Make Your Own Business To-Do List').toUpperCase()}</Title>
-        <CreateCategory />
-        {data && <CategoryDisplay categories={data} />}
-      </CategorySection>
-    </Container>
+    <CategorySection>
+      <Title>{String('Make Your Own Business To-Do List').toUpperCase()}</Title>
+      <CreateCategory />
+      {data && <CategoryDisplay categories={data} />}
+    </CategorySection>
   )
 }
