+++ client/app/ui/index.ts
@@ -1,6 +1,5 @@
 'use client'
 
 export * from './common'
-export * from './category'
 export * from './todolist'
 export * from './storage'
